[#5] Make project runnable: fixed typechecking problems

+++ src/modules/db/utils.py
@@ -54,5 +54,5 @@ async def cookie_file_ctx(
         else:
             yield None
     finally:
-        if cookie:
+        if cookie and cookie.file_path is not None:
             processing_utils.delete_file(cookie.file_path)
